backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
firebase-admin==6.2.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-jose==3.3.0
requests==2.31.0
django-environ>=0.11.2
google-auth>=2.25.0  # For Google authentication
django-filter>=23.5  # For filtering in DRF
drf-yasg>=1.21.7  # For API documentation
django-storages>=1.14  # For media file handling
django-cleanup>=8.0.0  # For automatic file cleanup 